+++ NEWS.md
@@ -12,6 +12,7 @@
 * Replace the `sample` function by a more solid one given by ?sample.
 * Removed unnecessary generic
 * Solver options and number of bootstrap resamples are now accessible as parameters to the Cort() function.
+* Add an option to force the checkerboard grid on trees and inside the forest.
 
 # cort 0.3.0